+++ tests/Spec.hs
@@ -6,6 +6,7 @@ import qualified Function.SizeSpec
 import qualified Function.MergeSpec
 import qualified Function.EachSpec
 import qualified Function.AssertPrivateSpec
+import qualified Function.JoinKeysToValuesSpec
 
 main :: IO ()
 main = hspec spec
@@ -17,7 +18,8 @@ spec = do
     describe "Class include" InterpreterSpec.classIncludeSpec
   describe "The shellquote function" Function.ShellquoteSpec.spec
   describe "stdlib functions" $ do
-      describe "The size function" Function.SizeSpec.spec
-      describe "The merge function" Function.MergeSpec.spec
-      describe "The each function" Function.EachSpec.spec
       describe "The assert_private function" Function.AssertPrivateSpec.spec
+      describe "The each function" Function.EachSpec.spec
+      describe "The join_keys_to_values function" Function.JoinKeysToValuesSpec.spec
+      describe "The merge function" Function.MergeSpec.spec
+      describe "The size function" Function.SizeSpec.spec